Split db_util.py into per-domain modules
db_util.py grew to 1224 lines mixing trigger, queue, job, log,
constant/credential and scheduler concerns. Split into trigger_util,
queue_util, job_util, log_util, settings_util and scheduler_util
(66-384 lines each). db_util keeps engine/session management (157
lines) and re-exports every function for backwards compatibility, so
existing db_util.X(...) callers — including the public
OrchestratorConnection SDK — work unchanged.
